Start migrating to vite
