+++ tucant/package.json
@@ -10,11 +10,19 @@
   },
   "dependencies": {
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.0",
+    "@mui/material": "^5.10.5",
+    "@mui/icons-material": "^5.10.3",
+    "@mui/lab": "^5.0.0-alpha.99",
+    "dompurify": "^2.4.0",
+    "@emotion/react": "^11.10.4",
+    "@emotion/styled": "^11.10.4"
   },
   "devDependencies": {
     "@types/react": "^18.0.21",
     "@types/react-dom": "^18.0.6",
+    "@types/dompurify": "^2.3.4",
     "@vitejs/plugin-react": "^2.1.0",
     "typescript": "^4.8.3",
     "vite": "^3.1.3"
