+++ tucant/src/App.tsx
@@ -21,7 +21,7 @@ function App() {
           count is {count}
         </button>
         <p>
-          Edit <code>src/App.tsx</code> and save to test HMR
+          Edit <code>src/App.tsx</code> and save to test HMR!
         </p>
       </div>
       <p className="read-the-docs">
