+++ tucant/src/assets/manifest.json
